backend/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile
#
annotated-types==0.7.0
    # via pydantic
anyio==4.7.0
    # via
    #   httpx
    #   starlette
    #   watchfiles
certifi==2024.8.30
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==1.17.1
    # via cryptography
charset-normalizer==3.4.0
    # via requests
click==8.1.7
    # via
    #   rich-toolkit
    #   typer
    #   uvicorn
cohere==5.13.3
    # via -r requirements.in
colorama==0.4.6
    # via
    #   click
    #   tqdm
    #   uvicorn
cryptography==44.0.0
    # via pyjwt
dnspython==2.7.0
    # via email-validator
docstring-parser==0.16
    # via llmeasytools
email-validator==2.2.0
    # via fastapi
fastapi[standard]==0.115.6
    # via -r requirements.in
fastapi-cli[standard]==0.0.6
    # via fastapi
fastavro==1.9.7
    # via cohere
filelock==3.16.1
    # via huggingface-hub
fsspec==2024.10.0
    # via huggingface-hub
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.7
    # via httpx
httptools==0.6.4
    # via uvicorn
httpx==0.28.1
    # via
    #   cohere
    #   fastapi
httpx-sse==0.4.0
    # via cohere
huggingface-hub==0.26.5
    # via tokenizers
idna==3.10
    # via
    #   anyio
    #   email-validator
    #   httpx
    #   requests
jinja2==3.1.4
    # via fastapi
llmeasytools==0.25
    # via -r requirements.in
markdown-it-py==3.0.0
    # via rich
markupsafe==3.0.2
    # via jinja2
mdurl==0.1.2
    # via markdown-it-py
numpy==1.26.4
    # via cohere
packaging==24.2
    # via huggingface-hub
parameterized==0.9.0
    # via cohere
pycparser==2.22
    # via cffi
pydantic==2.10.3
    # via
    #   cohere
    #   fastapi
    #   llmeasytools
pydantic-core==2.27.1
    # via
    #   cohere
    #   pydantic
pygments==2.18.0
    # via rich
pyjwt[crypto]==2.10.1
    # via -r requirements.in
python-dotenv==1.0.1
    # via
    #   -r requirements.in
    #   uvicorn
python-multipart==0.0.19
    # via fastapi
pyyaml==6.0.2
    # via
    #   huggingface-hub
    #   uvicorn
requests==2.32.3
    # via
    #   cohere
    #   huggingface-hub
rich==13.9.4
    # via
    #   rich-toolkit
    #   typer
rich-toolkit==0.12.0
    # via fastapi-cli
shellingham==1.5.4
    # via typer
sniffio==1.3.1
    # via anyio
starlette==0.41.3
    # via fastapi
tokenizers==0.21.0
    # via cohere
tqdm==4.67.1
    # via huggingface-hub
typer==0.15.1
    # via fastapi-cli
types-requests==2.32.0.20241016
    # via cohere
typing-extensions==4.12.2
    # via
    #   anyio
    #   cohere
    #   fastapi
    #   huggingface-hub
    #   pydantic
    #   pydantic-core
    #   rich-toolkit
    #   typer
urllib3==2.2.3
    # via
    #   requests
    #   types-requests
uvicorn[standard]==0.32.1
    # via
    #   fastapi
    #   fastapi-cli
watchfiles==1.0.0
    # via uvicorn
websockets==14.1
    # via uvicorn
